feat(UI): show username in appbar when scroll
