+++ lib/core/util/styles/profile_style.dart
@@ -5,7 +5,7 @@ import 'package:streamskit_mobile/core/util/themes/app_color.dart';
 const Color colorDividerBottomSheet = Color(0xFFC4C4C4);
 const Color colorDividerBottomSheetDark = Color(0xFF595959);
 
-///Text size: 13 color: mCL weight: 700
+/// Text size: 13 color: mCL weight: 700
 TextStyle text13w700mCL = TextStyle(
   color: mCL,
   fontSize: 13.sp,
@@ -13,27 +13,26 @@ TextStyle text13w700mCL = TextStyle(
   overflow: TextOverflow.ellipsis,
 );
 
-///Text size: 13 color: mCL
+/// Text size: 13 color: mCL
 TextStyle text13mCL = TextStyle(
   color: mCL,
   fontSize: 13.sp,
   overflow: TextOverflow.ellipsis,
 );
 
-///Text size: 11 color: mCL
+/// Text size: 11 color: mCL
 TextStyle text11mCL = TextStyle(
   color: mCL,
   fontSize: 11.sp,
   overflow: TextOverflow.ellipsis,
 );
 
-///Text size: 11 color:  colorBlack1
+/// Text size: 11 color:  colorBlack1
 TextStyle text11cB1 = TextStyle(
   color: colorBlack1,
   fontSize: 11.sp,
   overflow: TextOverflow.ellipsis,
 );
-
 
 ///Text size: 9 color: mCL
 TextStyle text9mCL = TextStyle(
